[ADD] 스킬 갯수 수정 전 커밋

+++ hackerton-ntb/src/main/java/com/ntb/hackertonntb/domain/repository/WantSkillRepository.java
@@ -2,10 +2,13 @@ package com.ntb.hackertonntb.domain.repository;
 
 import com.ntb.hackertonntb.domain.entity.WantSkill;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface WantSkillRepository extends JpaRepository<WantSkill, Integer> {
+public interface WantSkillRepository extends JpaRepository<WantSkill, Long> {
 
     List<WantSkill> findByIdContaining(String keyword);
+
 }
